Update 1_photoresistor-blink.ino
I propose the following because: 1. The original version will continuously publish the photoresistor status requiring a Hard Reset to reflash the Device.  2. Didn't include the on Board Led as the example annotation referenced and  3. Added Tinker Digital Write and Analog Read logic to use the Tinker app to test and operate the code...